requirements.txt
pyarrow>=21.0.0,<22.0.0
pyspark==4.1.1
python-dotenv>=1.1.1,<2.0.0
debugpy==1.8.16
pyyaml==6.0.3